requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
bitsandbytes>=0.41.0
peft>=0.8.0
datasets>=2.16.0
numpy>=1.24.0
opencv-python>=4.8.0
boto3>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2
aiofiles>=23.2.1
compel>=2.0.2
safetensors>=0.4.0
wandb>=0.16.0